backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

pydantic>=2.5.0
pydantic-settings>=2.1.0

motor>=3.3.2
pymongo>=4.6.0
redis>=5.0.1
celery>=5.3.4

ffmpeg-python>=0.2.0
opencv-python-headless>=4.8.1.78
pillow>=10.4.0

ultralytics>=8.0.206
fer>=25.10.0

torch==2.6.0
torchvision==0.21.0

git+https://github.com/openai/CLIP.git

langchain>=0.1.0
langgraph>=0.0.20
google-genai>=0.1.0
anthropic>=0.40.0

pyexiftool>=0.0.2

python-multipart>=0.0.12
websockets>=13.1
python-socketio>=5.11.0
aiohttp>=3.9.1

python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.1

httpx>=0.25.2
requests>=2.31.0

pytest>=7.4.3
pytest-asyncio>=0.21.1
ffmpeg-python